requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==0.24.0
async-generator==1.10
async-timeout==4.0.3
asyncssh==2.14.2
asynctest==0.13.0
atpublic==3.1.2
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
commonmark==0.9.1
configobj==5.0.8
contextlib2==21.6.0
contextvars==2.4
cryptography==42.0.5
cycler==0.11.0
databricks-cli==0.18.0
dataclasses==0.6
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==6.1.3
dpath==2.1.6
dulwich==0.21.7
dvc==2.10.2
dvc-render==0.0.5
dvclive==0.9.0
entrypoints==0.4
exceptiongroup==1.2.1
fastapi==0.103.2
fastjsonschema==2.19.1
Flask==2.2.5
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
funcy==2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
grandalf==0.6
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
idna-ssl==1.1.0
imageio==2.31.2
immutables==0.19
importlib-metadata==5.2.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab_widgets==3.0.10
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.39.1
mailchecker==6.0.3
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mlflow==1.30.1
msgpack==1.0.5
multidict==6.0.5
nanotime==0.5.2
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.6.3
notebook==6.5.6
notebook_shim==0.2.4
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.9.0
pexpect==4.9.0
phonenumbers==8.13.35
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
ply==3.11
pooch==1.8.1
prometheus-client==0.17.1
prometheus-flask-exporter==0.23.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycrypto==2.6.1
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
PyDrive2==1.15.4
pygit2==1.10.1
Pygments==2.17.2
PyGObject==3.26.1
pygtrie==2.5.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyrsistent==0.19.3
python-benedict==0.33.2
python-dateutil==2.9.0.post0
python-fsutil==0.14.1
python-slugify==8.0.4
pytz==2022.7.1
pyxdg==0.25
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
querystring-parser==1.2.4
requests==2.31.0
resampy==0.4.3
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.0.2
scipy==1.7.3
scmrepo==0.0.19
SecretStorage==2.3.1
Send2Trash==1.8.3
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
SQLAlchemy==1.4.52
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
torch==1.13.1
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.22.0
voluptuous==0.14.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.10
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.15.0
